src/backend/pyproject.toml
[project]
name = "tsuginote-backend"
version = "0.1.0"
description = "TsugiNote - AI-powered knowledge management platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "redis>=5.2.0",
    "stripe>=11.3.0",
    "openai>=1.58.0",
    "anthropic>=0.40.0",
    "qdrant-client>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"
